separated pages between Position Token and its details
