+++ src/App.tsx
@@ -3,7 +3,7 @@ import "focus-visible/dist/focus-visible"
 import Home from "./page/Home"
 import Trade from "./page/Trade"
 import PositionToken from "./page/PositionToken"
-import PositionTokenDetail from "./page/PositionToken/PositionTokenDetail"
+import PositionTokenDetail from "./page/PositionTokenDetail"
 import LiquidityProvider from "page/LiquidityProvider"
 import Pools from "page/Pools"
 import { Container, Divider, Box } from "@chakra-ui/react"
