+++ src/page/PositionToken/index.tsx
@@ -1,10 +1,18 @@
 import React from "react"
 import { Heading, Text } from "@chakra-ui/react"
 
+import { PerpdexLongTokenContainer } from "container/connection/perpdexLongTokenContainer"
+import { PerpdexMarketContainer } from "container/connection/perpdexMarketContainer"
+import { getLongTokenInfos } from "util/position"
 import FrameContainer from "component/FrameContainer"
 import TokenTable from "./TokenTable"
 
 function PositionToken() {
+    const { longTokenStates } = PerpdexLongTokenContainer.useContainer()
+    const { marketStates } = PerpdexMarketContainer.useContainer()
+
+    const longTokenInfos = getLongTokenInfos(longTokenStates, marketStates)
+
     return (
         <FrameContainer>
             <Heading size="lg" color="#627EEA">
@@ -15,7 +23,7 @@ function PositionToken() {
                 <br />
                 These tokens have the qualities such as liquidation free, and 100% capital efficient.
             </Text>
-            <TokenTable />
+            <TokenTable data={longTokenInfos} />
         </FrameContainer>
     )
 }
